Implemented component containing the action-buttons (edit and copy) of a link-card
